move contents of antilap and d2n to locfun root
