--- puncturedfem/locfun/antilap/__init__.py
@@ -1,14 +1,0 @@
-"""
-Anti-Laplacians of harmonic functions.
-
-Modules
--------
-antilap
-    Anti-Laplacian of a harmonic function.
-log_antilap
-    Anti-Laplacian of logarithmic harmonic function.
-"""
-
-from . import antilap, log_antilap
-
-__all__ = ["antilap", "log_antilap"]
