--- puncturedfem/locfun/d2n/__init__.py
@@ -1,20 +1,0 @@
-"""
-Dirichlet-to-Neumann map for harmonic functions.
-
-Modules
--------
-fft_deriv
-    Fourier differentiation.
-log_terms
-    Traces and derivatives of logarithmic terms.
-trace2tangential
-    Obtain a tangential derivative from a trace.
-"""
-
-from . import fft_deriv, log_terms, trace2tangential
-
-__all__ = [
-    "fft_deriv",
-    "log_terms",
-    "trace2tangential",
-]
